requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
google-genai==1.14.0
python-dotenv==1.1.0
httpx==0.28.1
faiss-cpu==1.11.0
sentence-transformers==4.1.0
qrcode[pil]==8.2
pydantic==2.11.3
pydantic[email]
motor==3.7.0
passlib[bcrypt]==1.7.4
PyJWT==2.10.1
